utom_feature/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.90
botocore==1.35.90
cassandra-driver==3.29.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
dramatiq>=1.15.0
elastic-transport==8.15.1
elasticsearch==8.16.0
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
geomet==0.2.1.post1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==7.34.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
jupyter-autotime==1.1.0
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.55.1
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pika>=1.3.2
platformdirs==4.3.6
playwright==1.49.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.2
pydantic_core==2.27.1
pyee==12.0.0
Pygments==2.18.0
pymongo>=4.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pyzmq==26.2.0
redis>=5.0.1
requests>=2.31.0
s3transfer==0.10.4
selenium>=4.16.0
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
pytube>=15.0.0
moviepy>=1.0.3
yt-dlp>=2023.11.16
undetected-chromedriver>=3.5.4
openai-whisper>=20231117
torch>=2.1.0
